refactor: Media
Update Media element
