+++ packages/ui-modules-media/MediaBody.tsx
@@ -1,13 +1,17 @@
-import React, { FC, HTMLAttributes } from 'react';
 import classNames from 'classnames';
+import React, { ReactNode } from 'react';
 import styles from '../media.css';
 
-export interface BodyProps extends HTMLAttributes<HTMLDivElement> {}
+interface MediaBodyProps {
+  children: ReactNode;
+  className?: string;
+}
 
-const MediaBody: FC<BodyProps> = (props) => {
+function MediaBody(props: MediaBodyProps): JSX.Element {
   const { className, ...otherProps } = props;
 
   return <div className={classNames(className, styles.media__body)} {...otherProps} />;
-};
+}
 
+export type { MediaBodyProps };
 export default MediaBody;
