--- packages/ui-modules-media/docs/MediaObject.stories.tsx
@@ -1,17 +1,0 @@
-import { Meta, Story } from '@storybook/react';
-import React from 'react';
-import { MediaObject } from '../react';
-import { MediaObjectProps } from '../react/MediaObject';
-
-export default {
-  component: MediaObject,
-  title: 'Media/MediaObject',
-} as Meta;
-
-const Template: Story<MediaObjectProps> = (args) => <MediaObject {...args} />;
-
-export const Default = Template.bind({});
-
-Default.args = {
-  direction: 'left',
-};
