+++ packages/ui-modules-media/index.ts
@@ -1,0 +1,3 @@
+import Media from './Media';
+
+export default Media;
