+++ packages/ui-modules-media/media.css
@@ -1,5 +1,6 @@
 .media {
   display: flex;
+  gap: 1em;
 }
 
 .media__object {
@@ -7,12 +8,10 @@
 }
 
 .media__object--left {
-  margin-right: 1em;
   order: -1;
 }
 
 .media__object--right {
-  margin-left: 1em;
   order: 1;
 }
