--- packages/ui-modules-media/react/index.ts
@@ -1,6 +1,0 @@
-import Media from './Media';
-import MediaBody from './MediaBody';
-import MediaObject from './MediaObject';
-
-export default Media;
-export { MediaBody, MediaObject };
