--- packages/ui-modules-media/react/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "index.js"
-}
